Enable accept header by default
The default value for the accept_header parameter in the async_impl client has been changed from false to true. This means the "accept: */*" header will now be inserted at the beginning by default. Changes have been made to both the parameter and its usage in the client builder function.